:truck: Move SRE stack components into 'stacks'
