+++ data_safe_haven/infrastructure/stacks/sre/application_gateway.py
