+++ data_safe_haven/infrastructure/stacks/sre/data.py
