+++ data_safe_haven/infrastructure/stacks/sre/hedgedoc_server.py
