+++ data_safe_haven/infrastructure/stacks/sre/monitoring.py
