+++ data_safe_haven/infrastructure/stacks/sre/networking.py
