+++ data_safe_haven/infrastructure/stacks/sre/remote_desktop.py
